+++ main_controller.py
@@ -1,5 +1,4 @@
 from main_model import MainModel
-from tasks_class import TasksForDay
 from user_class import User
 
 
@@ -8,15 +7,22 @@ class MainController:
         """
         :param _view: main window
         :param user: user object
-        :param tasks_for_today: class to get tasks for today
         """
 
         self._model = MainModel(user)
         self._view = _view
-        self.tasks_for_today = TasksForDay()
+
+    def get_tasks_for_day(self, date: str):
+        """
+        Get tasks for day
+        :param date: date in text format '%d/%m/%Y'
+        :return: tuple of tasks
+        """
+        tasks_for_today = self._model.try_to_get_tasks_for_given_date(date)
+        return tasks_for_today
 
     def submit_goal(self, data: str, date: str):
-        response = self._model.try_to_submit_goal(data, date)
+        response = self._model.try_to_submit_goal_for_given_date(data, date)
         return response
 
     def get_quote(self):
